requirements.txt
PyYAML==3.11
twilio==5.7.0
mysqlclient==1.3.9
python-daemon==2.1.2
